# packages/rest/src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"RecordId": {
				"type": "string",
				"example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
			},
...
			"ThreadId": {
				"type": "string",
				"example": "ea4e5e69-fc04-465a-90d2-9f8ff78aa71d"
			},
			"ProofState": {
				"description": "Present Proof protocol states as defined in RFC 0037",
				"enum": [
					"proposal-sent",
					"proposal-received",
					"request-sent",
					"request-received",
					"presentation-sent",
					"presentation-received",
					"declined",
					"abandoned",
					"done"
				],
				"type": "string"
			},
			"ProofRole": {
				"enum": [
					"verifier",
					"prover"
				],
				"type": "string"
			},
			"AutoAcceptProof": {
				"description": "Typing of the state for auto acceptance",
				"enum": [
					"always",
					"contentApproved",
					"never"
				],
				"type": "string"
			},
			"DidCommProofsExchangeRecord": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/RecordId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"type": "string"
					},
					"connectionId": {
						"$ref": "#/components/schemas/RecordId"
					},
					"threadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"parentThreadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"state": {
						"$ref": "#/components/schemas/ProofState"
					},
					"role": {
						"$ref": "#/components/schemas/ProofRole"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"errorMessage": {
						"type": "string"
					},
					"protocolVersion": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"type",
					"threadId",
					"state",
					"role",
					"protocolVersion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProofProtocolVersion": {
				"type": "string",
				"enum": [
					"v1",
					"v2"
				]
			},
...
			"AnonCredsProposeProofFormat": {
				"description": "Interface for creating an anoncreds proof proposal.",
				"properties": {
					"name": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"attributes": {
						"items": {
							"$ref": "#/components/schemas/AnonCredsPresentationPreviewAttribute"
						},
						"type": "array"
					},
					"predicates": {
						"items": {
							"$ref": "#/components/schemas/AnonCredsPresentationPreviewPredicate"
						},
						"type": "array"
					},
					"nonRevokedInterval": {
						"$ref": "#/components/schemas/AnonCredsNonRevokedInterval"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"DidCommProofsProposeProofOptions": {
				"properties": {
					"connectionId": {
						"type": "string"
					},
					"goalCode": {
						"type": "string"
					},
					"parentThreadId": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"protocolVersion": {
						"$ref": "#/components/schemas/ProofProtocolVersion"
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsProposeProofFormat"
							},
							"indy": {
								"$ref": "#/components/schemas/AnonCredsProposeProofFormat"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"connectionId",
					"protocolVersion",
					"proofFormats"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AcceptAnonCredsProposalOptions": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
...
			"DidCommProofsAcceptProposalOptions": {
				"properties": {
					"goalCode": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean",
						"default": true
					},
					"protocolVersion": {
						"$ref": "#/components/schemas/ProofProtocolVersion"
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AcceptAnonCredsProposalOptions"
							},
							"indy": {
								"$ref": "#/components/schemas/AcceptAnonCredsProposalOptions"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"protocolVersion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlaintextMessage": {
				"properties": {
					"@type": {
						"type": "string"
					},
					"@id": {
						"type": "string"
					},
					"~thread": {
						"properties": {
							"pthid": {
								"type": "string"
							},
							"thid": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"@type",
					"@id"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"DidCommProofsCreateRequestResponse": {
				"properties": {
					"message": {
						"$ref": "#/components/schemas/PlaintextMessage"
					},
					"proofExchange": {
						"$ref": "#/components/schemas/DidCommProofsExchangeRecord"
					}
				},
				"required": [
					"message",
					"proofExchange"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DidCommProofsCreateRequestOptions": {
				"properties": {
					"goalCode": {
						"type": "string"
					},
					"parentThreadId": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean",
						"default": true
					},
					"protocolVersion": {
						"$ref": "#/components/schemas/ProofProtocolVersion"
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsRequestProofFormatOptions"
							},
							"indy": {
								"$ref": "#/components/schemas/AnonCredsRequestProofFormatOptions"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"protocolVersion",
					"proofFormats"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidCommProofsSendRequestOptions": {
				"properties": {
					"goalCode": {
						"type": "string"
					},
					"parentThreadId": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean",
						"default": true
					},
					"protocolVersion": {
						"$ref": "#/components/schemas/ProofProtocolVersion"
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsRequestProofFormatOptions"
							},
							"indy": {
								"$ref": "#/components/schemas/AnonCredsRequestProofFormatOptions"
							}
						},
						"type": "object"
					},
					"connectionId": {
						"$ref": "#/components/schemas/RecordId"
					}
				},
				"required": [
					"protocolVersion",
					"proofFormats",
					"connectionId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AnonCredsSelectedCredentials": {
				"properties": {
					"attributes": {
						"$ref": "#/components/schemas/Record_string.AnonCredsRequestedAttributeMatch_"
					},
					"predicates": {
						"$ref": "#/components/schemas/Record_string.AnonCredsRequestedPredicateMatch_"
					},
					"selfAttestedAttributes": {
						"$ref": "#/components/schemas/Record_string.string_"
					}
				},
				"required": [
					"attributes",
					"predicates",
					"selfAttestedAttributes"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DidCommProofsAcceptRequestOptions": {
				"properties": {
					"goalCode": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean",
						"default": true
					},
					"useReturnRoute": {
						"type": "boolean",
						"description": "whether to enable return routing on the send presentation message. This value only\nhas an effect for connectionless exchanges."
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsSelectedCredentials"
							},
							"indy": {
								"$ref": "#/components/schemas/AnonCredsSelectedCredentials"
							}
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"AcceptCredentialProposalOptions": {
				"properties": {
					"credentialFormats": {
						"properties": {
							"indy": {
								"$ref": "#/components/schemas/AnonCredsAcceptProposalFormat"
							},
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsAcceptProposalFormat"
							}
						},
						"type": "object"
					},
					"autoAcceptCredential": {
						"$ref": "#/components/schemas/AutoAcceptCredential"
					},
					"comment": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"AcceptCredentialRequestOptions": {
				"properties": {
					"credentialFormats": {
						"$ref": "#/components/schemas/CredentialFormatPayload_CredentialFormats.acceptRequest_"
					},
					"autoAcceptCredential": {
						"$ref": "#/components/schemas/AutoAcceptCredential"
					},
					"comment": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"tenants": {
				"type": "apiKey",
				"name": "x-tenant-id",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "@credo-ts/rest",
		"version": "0.9.5",
		"description": "Rest API for using Credo over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/didcomm/proofs": {
			"get": {
				"operationId": "FindProofsByQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DidCommProofsExchangeRecord"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
												"protocolVersion": "v2",
												"role": "prover",
												"state": "proposal-sent",
												"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
												"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"autoAcceptProof": "always",
												"type": "ProofRecord"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Find proof exchanges by query",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "threadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					},
					{
						"in": "query",
						"name": "connectionId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ProofState"
						}
					},
					{
						"in": "query",
						"name": "parentThreadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ProofRole"
						}
					}
				]
			}
		},
		"/didcomm/proofs/{proofExchangeId}": {
			"get": {
				"operationId": "GetProofExchangeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidCommProofsExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieve proof exchange by proof exchange id",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteProof",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a proof exchange record.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
		"/didcomm/proofs/propose-proof": {
			"post": {
				"operationId": "ProposeProof",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidCommProofsExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Initiate a new presentation exchange as prover by sending a presentation proposal request\nto the connection with the specified connection id.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsProposeProofOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/proofs/{proofExchangeId}/accept-proposal": {
			"post": {
				"operationId": "AcceptProposal",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidCommProofsExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation proposal as verifier by sending an accept proposal message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsAcceptProposalOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/proofs/create-request": {
			"post": {
				"operationId": "CreateRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidCommProofsCreateRequestResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"proofExchange": {
												"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
												"protocolVersion": "v2",
												"role": "prover",
												"state": "proposal-sent",
												"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
												"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"autoAcceptProof": "always",
												"type": "ProofRecord"
											},
											"message": {
												"@id": "134b27f0-9366-4811-a36b-50bacfe57e61",
												"@type": "https://didcomm.org/present-proof/1.0/request-presentation"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates a presentation request not bound to any proposal or existing connection",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsCreateRequestOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/proofs/request-proof": {
			"post": {
				"operationId": "RequestProof",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates a presentation request bound to existing connection",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsSendRequestOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/proofs/{proofExchangeId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation request as prover by sending an accept request message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsAcceptRequestOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/proofs/{proofExchangeId}/accept-presentation": {
			"post": {
				"operationId": "AcceptPresentation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation as prover by sending an accept presentation message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
...
		"/didcomm/credentials/{credentialExchangeId}/accept-proposal": {
			"post": {
				"operationId": "AcceptProposal",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential proposal as issuer by sending an accept proposal message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialProposalOptions"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/credentials/{credentialExchangeId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential request as issuer by sending an accept request message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialRequestOptions"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/connections/{connectionId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"type": "ConnectionRecord",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
											"state": "invitation-sent",
											"role": "responder",
											"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
											"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a connection request as inviter by sending a connection response message\nfor the connection with the specified connection id.\n\nThis is not needed when auto accepting of connection is enabled.",
				"tags": [
					"DIDComm Connections"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}